VDP refactoring
