+++ src/Core/Snapshot.cs
@@ -54,7 +54,7 @@ public sealed class Snapshot : IEquatable<Snapshot>
   [Key(33)] public byte VScroll;
   [Key(34)] public byte HLineCounter;
   [Key(35)] public bool HLinePending;
-  [Key(36)] public bool ControlWritePending;
+  [Key(36)] public bool ControlWriteLatch;
   [Key(37)] public bool IRQ;
 
   // PSG
@@ -131,7 +131,7 @@ public sealed class Snapshot : IEquatable<Snapshot>
     if (VScroll              != other.VScroll)                  return false;
     if (HLineCounter         != other.HLineCounter)             return false;
     if (HLinePending         != other.HLinePending)             return false;
-    if (ControlWritePending  != other.ControlWritePending)      return false;
+    if (ControlWriteLatch   != other.ControlWriteLatch)      return false;
     if (IRQ                  != other.IRQ)                      return false;
     if (ChannelLatch         != other.ChannelLatch)             return false;
     if (VolumeLatch          != other.VolumeLatch)              return false;
